Refactoring: new Output namespace
