+++ src/Paraunit/Output/TestResult.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Paraunit\Output;
+
+/**
+ * Class TestResult
+ * @package Paraunit\Output
+ */
+class TestResult
+{
+
+}
