+++ src/Paraunit/Parser/OutputContainerBearerInterface.php
@@ -2,7 +2,7 @@
 
 namespace Paraunit\Parser;
 
-use Paraunit\Printer\OutputContainerInterface;
+use Paraunit\Output\OutputContainerInterface;
 
 /**
  * Interface OutputContainerBearerInterface
